requirements.txt
altgraph==0.17.4
black==25.1.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.1.8
contourpy==1.3.0
cycler==0.12.1
et_xmlfile==2.0.0
exceptiongroup==1.3.0
filelock==3.18.0
flake8==7.3.0
fonttools==4.59.0
fsspec==2025.7.0
hf-xet==1.1.5
huggingface-hub==0.33.4
idna==3.10
imageio==2.37.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
iniconfig==2.1.0
Jinja2==3.1.6
kiwisolver==1.4.7
lazy_loader==0.4
macholib==1.16.3
MarkupSafe==3.0.2
matplotlib==3.9.4
mpmath==1.3.0
mypy_extensions==1.1.0
networkx==3.2.1
numpy==2.0.2
opencv-python==4.12.0.88
openpyxl==3.1.5
packaging==25.0
pandas==2.3.1
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.3.8
pluggy==1.6.0
pooch==1.8.2
pycodestyle==2.14.0
pydicom==2.4.4
pyflakes==3.4.0
Pygments==2.19.2
pyinstaller==6.14.2
pyinstaller-hooks-contrib==2025.6
pylibjpeg==2.0.1
pylibjpeg-libjpeg==2.3.0
pylibjpeg-openjpeg==2.4.0
pyparsing==3.2.3
PyQt5==5.15.11
PyQt5-Qt5==5.15.17
PyQt5_sip==12.17.0
PyQt6==6.9.1
PyQt6-Qt6==6.9.1
PyQt6_sip==13.10.2
pyqtgraph==0.13.7
pytest==8.4.1
python-dateutil==2.9.0.post0
python-gdcm==3.0.26
pytz==2025.2
PyYAML==6.0.2
reportlab==4.4.2
requests==2.32.4
safetensors==0.5.3
scikit-image==0.24.0
scipy==1.13.1
segmentation_models_pytorch==0.5.0
simpleitk==2.5.2
six==1.17.0
sympy==1.14.0
tifffile==2024.8.30
timm==1.0.17
tomli==2.2.1
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
zipp==3.23.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
pre-commit==3.5.0
